+++ CHANGELOG.md
@@ -7,7 +7,13 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
 - Remove archived games from the `ron` file.
+
+### Fixed
+
+- hnefatafl-server-full: don't send empty strings.
 
 ## [0.7.0] - 2025-02-23